add memoization to few validators
