+++ src/routes/safe/components/Balances/Tokens/screens/AddCustomToken/validators.js
@@ -3,19 +3,8 @@ import { List } from 'immutable'
 import { getWeb3 } from '~/logic/wallets/getWeb3'
 import { type Token } from '~/logic/tokens/store/model/token'
 import { sameAddress } from '~/logic/wallets/ethAddresses'
+import { simpleMemoize } from '~/components/forms/validator'
 // import { getStandardTokenContract } from '~/logic/tokens/store/actions/fetchTokens'
-
-export const simpleMemoize = (fn: Function) => {
-  let lastArg
-  let lastResult
-  return (arg: any) => {
-    if (arg !== lastArg) {
-      lastArg = arg
-      lastResult = fn(arg)
-    }
-    return lastResult
-  }
-}
 
 // eslint-disable-next-line
 export const addressIsTokenContract = simpleMemoize(async (tokenAddress: string) => {
